Integracion de redux
