+++ netcli/package.json
@@ -11,12 +11,15 @@
   "dependencies": {
     "@headlessui/react": "^1.7.13",
     "@heroicons/react": "^2.0.16",
+    "@reduxjs/toolkit": "^1.9.3",
     "axios": "^1.3.3",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
     "react-router": "^6.8.1",
-    "react-router-dom": "^6.8.1"
+    "react-router-dom": "^6.8.1",
+    "redux": "^4.2.1"
   },
   "devDependencies": {
     "@types/react": "^18.0.27",
