+++ netcli/src/components/VerPelicula.jsx
@@ -2,10 +2,15 @@ import { Link, useLocation } from "react-router-dom";
 import { Disclosure } from "@headlessui/react";
 import { ChevronUpIcon } from "@heroicons/react/20/solid";
 import { useEffect, useContext, useState } from "react";
-import { Contexto_Funciones } from "../context/contextoFunciones";
+import { useDispatch, useSelector } from "react-redux";
+import { moviesList } from "../features/movies/moviesSlice";
+import { movieList } from "../features/movies/movieSlice";
+import http from "../http-common"
 
 export default function VerPelicula() {
-	const { peliSelect, peliculaSelect, activo, crearReview, eliminarPeli } = useContext(Contexto_Funciones);
+	const dispatch = useDispatch()
+	const activo = useSelector(state => state.activo)
+	const peliSelect = useSelector(state => state.movie)
   const [review, setReview] = useState('')
   const [valor, setValor] = useState(1)
 	function textareaHeight(textarea) {
@@ -22,19 +27,46 @@ export default function VerPelicula() {
 			setValor(1)
 		}
 	}
+	function eliminarPeli(mov, img) {
+		http
+		.delete(`/eliminar/${mov}/${img}`)
+		.then((response) => {
+			console.log(response.data)
+			dispatch(moviesList())
+			window.location.href = '/';
+		})
+		.catch((error) => {
+			console.log('Error:', error);
+			// Alert('Ha sucedido un problema', false)
+		});
+	}
+  function crearReview(user, mov, review, valor) {
+    if (review && valor) {
+      http
+      .post("/aggreview", {user: user, mov: mov, review: review, valor: valor})
+      .then((response) => {
+        console.log(response.data)
+        dispatch(movieList(mov))
+      })
+      .catch((error) => {
+        console.log('Error:', error);
+        // Alert('Ha sucedido un problema', false)
+      });
+    }
+  }
 	const location = useLocation()
 	const route = location.pathname.split("/")[2];
 	if (route) {
 		useEffect((e) => {
-			peliculaSelect(route)	
+			dispatch(movieList(route))
 		}, []);
-		if (peliSelect.id) {
+		if (peliSelect.mov.id) {
 			return (
 				<>
 					<div className="mt-5">
 						<iframe
 							className="w-full h-[20rem] sm:h-[30rem] lg:h-[40rem]"
-							src={peliSelect.link}
+							src={peliSelect.mov.link}
 							title="YouTube video player"
 							allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
 							allowFullScreen
@@ -45,9 +77,9 @@ export default function VerPelicula() {
 									{({ open }) => (
 										<>
 											<Disclosure.Button className="flex w-full justify-between items-center rounded-lg px-4 py-2 text-left text-sm font-medium focus:outline-none focus-visible:ring focus-visible:ring-primario focus-visible:ring-opacity-75">
-												<button type="button" onClick={(e) => eliminarPeli(peliSelect.id, peliculaSelect.img)}>Eliminar</button>
+												<button type="button" onClick={(e) => eliminarPeli(peliSelect.mov.id, peliSelect.mov.img)}>Eliminar</button>
 												<div>
-													<h3 className="text-3xl">{peliSelect.head}</h3>
+													<h3 className="text-3xl">{peliSelect.mov.head}</h3>
 													<p className="text-sm">Ver detalles</p>
 												</div>
 												<ChevronUpIcon
@@ -60,13 +92,13 @@ export default function VerPelicula() {
 														<div className="bg-stone-light px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
 															<dt className="text-sm font-bold text-stone-dark">Titulo</dt>
 															<dd className="mt-1 text-sm sm:col-span-2 sm:mt-0">
-																{peliSelect.head}
+																{peliSelect.mov.head}
 															</dd>
 														</div>
 														<div className="bg-white px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
 															<dt className="text-sm font-bold text-stone-dark">Director</dt>
 															<dd className="mt-1 text-sm sm:col-span-2 sm:mt-0">
-															{peliSelect.director}
+															{peliSelect.mov.director}
 															</dd>
 														</div>
 														<div className="bg-stone-light px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
@@ -74,13 +106,13 @@ export default function VerPelicula() {
 																Sipnosis
 															</dt>
 															<dd className="mt-1 text-sm sm:col-span-2 sm:mt-0">
-																{peliSelect.sipnosis}
+																{peliSelect.mov.sipnosis}
 															</dd>
 														</div>
 														<div className="bg-white px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
 															<dt className="text-sm font-bold text-stone-dark">Franquicia</dt>
 															<dd className="mt-1 text-sm sm:col-span-2 sm:mt-0">
-															{peliSelect.franquicia}
+															{peliSelect.mov.franquicia}
 															</dd>
 														</div>
 														<div className="bg-stone-light px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
@@ -88,21 +120,21 @@ export default function VerPelicula() {
 																Genero
 															</dt>
 															<dd className="mt-1 text-sm sm:col-span-2 sm:mt-0">
-																{peliSelect.genero}
+																{peliSelect.mov.genero}
 															</dd>
 														</div>
 														<div className="bg-white px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
 															<dt className="text-sm font-bold text-stone-dark">
 																Fecha de publicación
 															</dt>
-															<dd className="mt-1 text-sm sm:col-span-2 sm:mt-0">{peliSelect.fecMov}</dd>
+															<dd className="mt-1 text-sm sm:col-span-2 sm:mt-0">{peliSelect.mov.fecMov}</dd>
 														</div>
 														<div className="bg-stone-light px-4 py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
 															<dt className="text-sm font-bold text-stone-dark">
 																Actores
 															</dt>
 															<dd className="mt-1 text-sm sm:col-span-2 sm:mt-0">
-																{peliSelect.actors}
+																{peliSelect.mov.actors}
 															</dd>
 														</div>
 													</dl>
@@ -113,7 +145,7 @@ export default function VerPelicula() {
 								</Disclosure>
 								<p className="mt-4">Comentarios</p>
 								<ul>
-									{peliSelect.reviews.map((comentario) => (
+									{peliSelect.mov.reviews.map((comentario) => (
 										<li key={comentario.id} className="my-6 divide-y divide-muted-neutral">
 											<div className="flex gap-x-6">
 												<img
@@ -159,7 +191,7 @@ export default function VerPelicula() {
 											</div>
 											<button 
 												onClick={(e) => {
-													crearReview(activo.user.idUser, peliSelect.id, review, valor);
+													crearReview(activo.user.idUser, peliSelect.mov.id, review, valor);
 													limpiarCampos()
 												}}
 											> Enviar </button>
@@ -172,7 +204,7 @@ export default function VerPelicula() {
 									<div className="mt-8">
 										<div className="flow-root">
 											<ul role="list" className="my-6 divide-y divide-muted-neutral">
-												{peliSelect.recomendacion.map((pelicula) => (
+												{peliSelect.mov.recomendacion.map((pelicula) => (
 													<li
 														key={pelicula.idMov}
 														className="flex flex-col md:flex-row lg:flex-col xl:flex-row items-center py-6"
