+++ netcli/src/main.jsx
@@ -1,13 +1,14 @@
-import { Contexto_DataProvider } from './context/contextoFunciones';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './styles/index.css';
+import { Provider } from 'react-redux'
+import {store} from './app/store.js'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <Contexto_DataProvider>
+    <Provider store={store}>
       <App />
-    </Contexto_DataProvider>
+    </Provider>
   </React.StrictMode>
 );
